change to integer year and month

+++ statWebApp/encounter/models.py
@@ -10,7 +10,7 @@ class Disease(models.Model):
 
 
 class Encounter(models.Model):
-	age = models.FloatField(null=True)
+	age = models.IntegerField(null=True)
 	ageMonth = models.IntegerField(null=True)
 	gender = models.CharField(max_length = 10, null=True)
 	priorPatientStatus = models.CharField(max_length = 12, null=True)
